docs: update docs and stories for first 6 data-entry components
