+++ docs/components/Data Entry/Auto Complete/Documentation.mdx
@@ -8,8 +8,16 @@ import * as AutoCompleteStories from '../../../../src/components/data-entry/Auto
 
 # AutoComplete
 
-This is the documentation for the AutoComplete component
+#### Overview
 
-{/* Documentation goes here */}
+An input field with text hints that aids users by providing keyword-based suggestions as they type.
+
+#### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [AutoComplete Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?m=auto&node-id=388-11035&t=zURbLpG60aM6aJiH-1) |
+| AntD | [AutoComplete Component](https://ant.design/components/auto-complete) |
+
 
 <Canvas meta={AutoCompleteStories.default} />
