+++ docs/components/Data Entry/Checkbox/Documentation.mdx
@@ -8,8 +8,27 @@ import * as CheckboxStories from '../../../../src/components/data-entry/Checkbox
 
 # Checkbox
 
-This is the documentation for the checkbox component
+#### Overview
 
-{/* Documentation goes here */}
+The **Checkbox** component allows users to select one or multiple options from a list of choices. It provides a straightforward way to toggle selections on or off and can be used individually or within a group of related options.
+
+### When to Use
+
+Use the **Checkbox** component when:
+- In lists or forms where users can select multiple items.
+- For binary options, especially when changes require saving.
+- For active agreements, such as accepting terms of service.
+
+When selecting a single option from a set of mutually exclusive choices, use **[Radio Buttons](https://mparticle.github.io/aquarium/?path=/story/components-data-entry-radio--primary)**.
+
+Use **[switch](https://mparticle.github.io/aquarium/?path=/story/components-data-entry-switch--primary)** for cases like activation, filter controls, or subscriptions where there is a clear "on/off" state.
+
+#### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Checkbox Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=388-11036&p=f&t=zURbLpG60aM6aJiH-0) |
+| AntD | [Checkbox Component](https://ant.design/components/checkbox) |
+
 
 <Canvas meta={CheckboxStories.default} />
