+++ docs/components/Data Entry/Date Picker/Documentation.mdx
@@ -8,8 +8,15 @@ import * as DatePickerStories from '../../../../src/components/data-entry/DatePi
 
 # Date Picker
 
-This is the documentation for the date picker component
+#### Overview
+Used to select or input a single date through an interactive popup calendar.
+For date range selection with presets, use the [Date Range Filter](https://mparticle.github.io/aquarium/?path=/story/candidate-components-directory-date-range-filter--date-range-filter).
 
-{/* Documentation goes here */}
+#### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [DatePicker Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=388-11437&p=f&t=zURbLpG60aM6aJiH-0) |
+| AntD | [DatePicker Component](https://ant.design/components/date-picker) |
 
 <Canvas meta={DatePickerStories.default} />
