+++ docs/components/Data Entry/Input/Documentation.mdx
@@ -8,8 +8,19 @@ import * as InputStories from '../../../../src/components/data-entry/Input/Input
 
 # Input
 
-This is the documentation for the Inputs component
+#### Overview
+The **Input** component allows users to enter content using a mouse or keyboard. It serves as the most basic form field wrapper and is commonly used where user input is required.
 
-{/* Documentation goes here */}
+#### **When To Use**
+- A user input in a form field is needed.
+- A search input is required.
+
+#### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Input Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=388-11439&p=f&t=zURbLpG60aM6aJiH-0) |
+| AntD | [Input Component](https://ant.design/components/input) |
+
 
 <Canvas meta={InputStories.default} />
