+++ docs/components/Data Entry/InputNumber/Documentation.mdx
@@ -8,8 +8,18 @@ import * as InputNumberStories from '../../../../src/components/data-entry/Input
 
 # InputNumber
 
-This is the documentation for the InputNumbers component
+#### Overview
+The **InputNumber** component allows users to enter numeric values within a specified range using the mouse or keyboard. It ensures controlled and validated number input.
 
-{/* Documentation goes here */}
+#### **When To Use**
+- When a numeric value needs to be provided by the user.
+- To restrict input to numbers only, optionally within a defined range.
+
+#### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [InputNumber Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=388-11439&p=f&t=zURbLpG60aM6aJiH-0) |
+| AntD | [InputNumber Component](https://ant.design/components/input-number) |
 
 <Canvas meta={InputNumberStories.default} />
