+++ src/components/data-entry/Cascader/Cascader.stories.tsx
@@ -42,7 +42,7 @@ const options: ICascaderProps['options'] = [
 ]
 
 const meta: Meta<typeof Cascader> = {
-  title: 'Components/Data Entry/Cascader',
+  title: 'Components/Not Prod Ready/Data Entry/Cascader',
   component: Cascader,
 
   args: {
